refactor: Move spacing to theme
